+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- `is-invalid` added to invalid input groups to show form feedback
+
 ## [1.0.1] - 2023-07-18
 
 ### Changed